Refactor and reorganize project structure
