+++ alembic/env.py
@@ -5,7 +5,7 @@ from sqlalchemy import pool
 
 from alembic import context
 
-import models
+import app.models.models as models
 
 # changed in p86
 # this is the Alembic Config object, which provides
