+++ app/api/analytics.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, HTTPException, status, Depends
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
-from database import SessionLocal
+from app.database.database import SessionLocal
 
 router = APIRouter(
 	prefix="/analytics",
