+++ app/utils/barcode_utils.py
