+++ app/utils/qrcode_utils.py
